fix all belum si keuangan

+++ database/migrations/2022_04_11_143558_create_tagihan_pelaksanaan_pekerjaan_table.php
@@ -17,7 +17,7 @@ class CreateTagihanPelaksanaanPekerjaanTable extends Migration
             //FOREIGN KEY CONSTRAINTS
             $table->string('tagihan_id')->references('id')->on('tagihan')->onDelete('cascade');
             $table->string('pelaksanaan_pekerjaan_id')->references('id')->on('pelaksanaan_pekerjaan')->onDelete('cascade');
-            $table->integer('total')->default(0);
+            $table->float('total', 20, 3)->default(0);
             //SETTING THE PRIMARY KEYS
             $table->primary(['tagihan_id', 'pelaksanaan_pekerjaan_id']);
             $table->timestamps();
